requirements.txt
transformers
datasets
sentence-transformers
faiss-cpu
streamlit
ragas
openai